[qwen3.5] Separate GDN modules from model
