+++ torchtitan/experiments/rl/controller.py
@@ -411,24 +411,6 @@ class Controller(Configurable):
                     "pull reuse KV cached under the old weights."
                 )
 
-            # FULL cudagraph is only correct with the flex attention backend
-            cudagraph = self.generator.cudagraph
-            if (
-                cudagraph.enable
-                and cudagraph.mode == "FULL"
-                and self.model_spec is not None
-            ):
-                from torchtitan.models.common.attention import FlexAttention
-
-                inner_attn = self.model_spec.model.first_inner_attention
-                if not isinstance(inner_attn, FlexAttention.Config):
-                    raise ValueError(
-                        "cudagraph mode 'FULL' is only supported with the flex "
-                        "attention backend; the varlen backend corrupts FULL capture "
-                        "of mixed prefill+decode batches (#3709). Use FULL_DECODE_ONLY "
-                        "or FULL_AND_PIECEWISE."
-                    )
-
     def __init__(self, config: Config):
         self.config = config
         self.trainer: PolicyTrainer | None = None
