+++ torchtitan/models/qwen3_5/__init__.py
@@ -38,17 +38,14 @@ from torchtitan.protocols.model import ModelConfigConverter
 
 from torchtitan.protocols.model_spec import ModelSpec
 
-from .model import (
+from .gdn import (
     GatedDeltaBackend,
     GatedDeltaKernel,
     GatedDeltaNet,
-    GatedDeltaNetCore,
-    OffsetRMSNorm,
-    Qwen35Attention,
-    Qwen35Model,
-    Qwen35TransformerBlock,
+    InnerGatedDeltaNet,
     RMSNormGated,
 )
+from .model import OffsetRMSNorm, Qwen35Attention, Qwen35Model, Qwen35TransformerBlock
 
 from .parallelize import parallelize_qwen3_5
 from .rope import MRoPE
@@ -304,7 +301,7 @@ def _qwen35_deltanet_config(
         conv_q=_conv(key_dim),
         conv_k=_conv(key_dim),
         conv_v=_conv(value_dim),
-        core=GatedDeltaNetCore.Config(
+        inner_gated_delta_net=InnerGatedDeltaNet.Config(
             key_head_dim=key_head_dim,
             value_head_dim=value_head_dim,
             key_dim=key_dim,
